Basic ICS parser
More a proof-of-concept than anything else.  I want this eventually to
be able to parse the .ics files that I get in emails into REM lines
(which the ics2rem I've found in the past does not do).

Ideally, it may be able to pop up a dialog to allow the receiver of an
ICS file to edit the REM line and append it to a file of their
choosing; even configured so that it will automatically append to a
specific file.